DataIngress/requirements.txt
snowflake-connector-python
sqlalchemy
psycopg2-binary
pandas
python-dotenv
tqdm
